extract dhis logic for event edition
