+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventCapture/EventCaptureContract.java
@@ -158,8 +158,6 @@ public class EventCaptureContract {
 
         boolean isEventEditable(String eventUid);
 
-        boolean optionIsInOptionGroup(String optionUid, String optionGroupToHide);
-
         String getSectionFor(String field);
 
         Single<Boolean> canReOpenEvent();
